Change primitive data type

+++ Week #2/Calculation VAT/src/Main.java
@@ -3,14 +3,14 @@ public class Main {
     public static void main(String[] args){
 
         //Variables are defined
-        int price,vat,priceExVAT;
+        double price,vat,priceExVAT;
 
         //Scanner class was defined
         System.out.println("Please Enter a Price of Product");
         Scanner input = new Scanner(System.in);
 
         //The value obtained from user is assigned a variable
-        price = input.nextInt();
+        price = input.nextDouble();
 
         priceExVAT = price - ((price/100) * 18);
